Extract field ranking helpers into dedicated module
